fix: 情绪指数小数点问题 - 所有分数显示强制取整
- market-emotion-api.py: total_score 使用 int(round()) 确保返回整数
- index.html: 主分数显示加 Math.round()
- index.html: 维度分数显示加 Math.round() (弹窗 + 主界面)
- index.html: 模拟数据更新时取整
- 添加 FIX_REPORT 修复报告文档

修复问题：情绪指数显示 69.20108379223133 而非整数 69

+++ market-emotion-api.py
@@ -148,7 +148,7 @@ def get_market_emotion():
         scores = calculate_scores(volume, up_down_ratio, leader_height, theme_change, index_change)
         
         # 计算综合分数
-        total_score = round(sum(scores.values()) / 5)
+        total_score = int(round(sum(scores.values()) / 5))
         
         # 确定状态
         if total_score <= 20:
